+++ packages/core/src/components.d.ts
@@ -167,6 +167,11 @@ export namespace Components {
         "type": 'checkbox' | 'radio';
         "value": string;
     }
+    interface DsoToggletip {
+        "label": string;
+        "position": "top" | "right" | "bottom" | "left";
+        "small"?: boolean;
+    }
     interface DsoTooltip {
         /**
           * Activate the tooltip (Sets the `active` attribute)
@@ -196,6 +201,10 @@ export namespace Components {
           * Defines if the tooltip has a smaller max-width
          */
         "small"?: boolean;
+        /**
+          * Deactivates mouseover behaviour
+         */
+        "stateless"?: boolean;
     }
 }
 declare global {
@@ -313,6 +322,12 @@ declare global {
         prototype: HTMLDsoSelectableElement;
         new (): HTMLDsoSelectableElement;
     };
+    interface HTMLDsoToggletipElement extends Components.DsoToggletip, HTMLStencilElement {
+    }
+    var HTMLDsoToggletipElement: {
+        prototype: HTMLDsoToggletipElement;
+        new (): HTMLDsoToggletipElement;
+    };
     interface HTMLDsoTooltipElement extends Components.DsoTooltip, HTMLStencilElement {
     }
     var HTMLDsoTooltipElement: {
@@ -339,6 +354,7 @@ declare global {
         "dso-progress-bar": HTMLDsoProgressBarElement;
         "dso-progress-indicator": HTMLDsoProgressIndicatorElement;
         "dso-selectable": HTMLDsoSelectableElement;
+        "dso-toggletip": HTMLDsoToggletipElement;
         "dso-tooltip": HTMLDsoTooltipElement;
     }
 }
@@ -514,6 +530,11 @@ declare namespace LocalJSX {
         "type": 'checkbox' | 'radio';
         "value": string;
     }
+    interface DsoToggletip {
+        "label"?: string;
+        "position"?: "top" | "right" | "bottom" | "left";
+        "small"?: boolean;
+    }
     interface DsoTooltip {
         /**
           * Whether or not to show the tooltip. To control the tooltip add the `active` attribute or use the `activate()` and `deactivate()` instance methods.
@@ -535,6 +556,10 @@ declare namespace LocalJSX {
           * Defines if the tooltip has a smaller max-width
          */
         "small"?: boolean;
+        /**
+          * Deactivates mouseover behaviour
+         */
+        "stateless"?: boolean;
     }
     interface IntrinsicElements {
         "dso-alert": DsoAlert;
@@ -556,6 +581,7 @@ declare namespace LocalJSX {
         "dso-progress-bar": DsoProgressBar;
         "dso-progress-indicator": DsoProgressIndicator;
         "dso-selectable": DsoSelectable;
+        "dso-toggletip": DsoToggletip;
         "dso-tooltip": DsoTooltip;
     }
 }
@@ -582,6 +608,7 @@ declare module "@stencil/core" {
             "dso-progress-bar": LocalJSX.DsoProgressBar & JSXBase.HTMLAttributes<HTMLDsoProgressBarElement>;
             "dso-progress-indicator": LocalJSX.DsoProgressIndicator & JSXBase.HTMLAttributes<HTMLDsoProgressIndicatorElement>;
             "dso-selectable": LocalJSX.DsoSelectable & JSXBase.HTMLAttributes<HTMLDsoSelectableElement>;
+            "dso-toggletip": LocalJSX.DsoToggletip & JSXBase.HTMLAttributes<HTMLDsoToggletipElement>;
             "dso-tooltip": LocalJSX.DsoTooltip & JSXBase.HTMLAttributes<HTMLDsoTooltipElement>;
         }
     }