+++ packages/sources/src/components/toggletip/toggletip.models.ts
@@ -1,0 +1,8 @@
+import { tooltipPositions } from "../tooltip/tooltip.models";
+
+export interface Toggletip {
+  children: string;
+  position?: typeof tooltipPositions;
+  small?: boolean;
+  label?: string;
+}